refactor(material): D-410 物料到货/入库/库存/对账数量支持小数（INT → DECIMAL 12,4）
面料按米/公斤/码计量时数量是小数（如采购 1.32 米、到货 1.32 米），
但四个字段仍是 INT，前端输入框支持两位小数、后端用 int 接收 → 1.32 被静默截断成 1。

后果（P0 财务口径）：对账单按「到货量 × 单价」结算，截断后 1×60=60，
真实应付 1.32×60=79.20 → 少付货款；且采购单打印显示 79.20，两边对不上。

改动字段（配套迁移 V202709150100，幂等）：
- t_material_purchase.arrived_quantity
- t_material_inbound.inbound_quantity
- t_material_stock.quantity（入库 1.32 米库存只加 1 → 账实不符）
- t_material_reconciliation.quantity

实现要点：
- 比较一律改 compareTo(BigDecimal.ZERO)，三元默认值 0 → BigDecimal.ZERO
- 加减改 add/subtract，日志 %d → %s + stripTrailingZeros()
- 防重总闸改用 BigDecimal，否则「已入库 1 + 本次 1.32 > 1」会误拦小数物料
- MaterialStockService 加 decreaseStockById/decreaseStockForCancelReceive 的
  BigDecimal 重载（int 版委托过去），mapper 因 MyBatis 不能重载而另起
  decreaseStockWithCheckDecimal
- 仍为 Integer 的字段（领料单/出库单/告警 DTO 等，未纳入本次迁移）按原语义取整，
  并在代码注释标注，避免误以为已全链路支持小数

+++ backend/src/main/java/com/fashion/supplychain/finance/entity/MaterialReconciliation.java
@@ -86,7 +86,11 @@ public class MaterialReconciliation implements com.fashion.supplychain.finance.s
     @TableField(exist = false)
     private Integer productionCompletedQuantity;
 
-    private Integer quantity;
+    /**
+     * 对账数量（D-410：INT → BigDecimal，DECIMAL(12,4)）
+     * 对账单金额 = 单价 × 本数量，INT 会把 1.32 米截断成 1 → 少付货款。
+     */
+    private BigDecimal quantity;
 
     private BigDecimal unitPrice;
